+++ src/renderer/components/node/CollapsedHandles.tsx
@@ -1,19 +1,14 @@
 import { NeverType } from '@chainner/navi';
 import { Box } from '@chakra-ui/react';
-import { memo, useMemo } from 'react';
-import { Handle, Node, Position, useReactFlow } from 'reactflow';
-import { useContext, useContextSelector } from 'use-context-selector';
-import { InputId, NodeData } from '../../../common/common-types';
-import {
-    parseSourceHandle,
-    parseTargetHandle,
-    stringifySourceHandle,
-    stringifyTargetHandle,
-} from '../../../common/util';
+import { memo } from 'react';
+import { Handle, Position } from 'reactflow';
+import { useContextSelector } from 'use-context-selector';
+import { InputId } from '../../../common/common-types';
+import { stringifySourceHandle, stringifyTargetHandle } from '../../../common/util';
 import { BackendContext } from '../../contexts/BackendContext';
-import { GlobalVolatileContext } from '../../contexts/GlobalNodeState';
 import { defaultColor, getTypeAccentColors } from '../../helpers/accentColors';
 import { NodeState } from '../../helpers/nodeState';
+import { useSourceTypeColor } from '../../hooks/useSourceTypeColor';
 import { getBackground } from '../Handle';
 
 interface InputHandleProps {
@@ -24,39 +19,7 @@ interface InputHandleProps {
 }
 
 const InputHandle = memo(({ nodeId, isIterated, inputId, handleId }: InputHandleProps) => {
-    const { edgeChanges, typeState } = useContext(GlobalVolatileContext);
-    const { getEdges, getNode } = useReactFlow();
-
-    const connectedEdge = useMemo(() => {
-        return getEdges().find(
-            (e) => e.target === nodeId && parseTargetHandle(e.targetHandle!).inputId === inputId
-        );
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [edgeChanges, getEdges, nodeId, inputId]);
-
-    const { functionDefinitions } = useContext(BackendContext);
-
-    const sourceTypeColor = useMemo(() => {
-        if (connectedEdge) {
-            const sourceNode: Node<NodeData> | undefined = getNode(connectedEdge.source);
-            const sourceOutputId = parseSourceHandle(connectedEdge.sourceHandle!).outputId;
-            if (sourceNode) {
-                const sourceDef = functionDefinitions.get(sourceNode.data.schemaId);
-                if (!sourceDef) {
-                    return defaultColor;
-                }
-                const sourceType =
-                    typeState.functions.get(sourceNode.id)?.outputs.get(sourceOutputId) ??
-                    sourceDef.outputDefaults.get(sourceOutputId);
-                if (!sourceType) {
-                    return defaultColor;
-                }
-                return getTypeAccentColors(sourceType)[0];
-            }
-            return defaultColor;
-        }
-        return null;
-    }, [connectedEdge, functionDefinitions, typeState, getNode]);
+    const sourceTypeColor = useSourceTypeColor(nodeId, inputId);
 
     return (
         <Box
